+++ lib/tdc/generators/configurable_generator.rb
@@ -23,7 +23,7 @@ module Tdc
       # Hook method: subclasses are expected to define how to generate a model instance.
       #
       def generate_instance
-        raise Tdc::MissingOverrideError
+        raise Tdc::MissingOverrideError, "Implement the 'generate_instance' method"
       end
 
       #